scripts: add JSON Schema for device metadata
> JSON Schema is a vocabulary that allows you to annotate and validate
JSON documents.

Adding new devices to OpenWrt includes traditionally an arbitrary
formatted description of the device. There are no specifications for the
exact format, which makes it generic parsing impossible. The metadata is
added in an additional step to the OpenWrt Wiki.

Instead of storing device information in the Git commit log, the
metadata information could also be added in a machine readable format.

YAML is an established format for storing data, however it is complex
enough to allow badly formatted documents. For automatic validation of
created documents the "JSON Schema" can be used, which confusingly
contains the word JSON, while it really is a vocabulary of rules to
describe allowed contents.

The added "JSON schema" definitions allow the validation of device
metadata files. These files describe devices similar to the techdata
found in the OpenWrt wiki, however only contain static information which
does not change over time.

Signed-off-by: Paul Spooren <mail@aparcar.org>